Summary - Added a basic abstract Response class, and created a working definition for a SimpleMediaType based on JSON.
Controller.php: Removed dependency on Header.php as that functionality now is within the Response class
Headers.php: Removed file and placed relevant code within Response class
Response.php: New abstract class to act as a base for response objects, it includes the ability to setup a header based on a status code.
SimpleMediaType: An attempt to define a simple JSON based media type for use in early development, does not fully support the data model and is not to be used in production code.