Summary - Removed ability of Node and Link to construct with a NULL DFD, made Links and Nodes save themselves to DFDs
DFDModel/Link.php: Removed the portion of the constructor that allowed
null DFDs as a parent.
DFDModel/Node.php: Removed the portion of the constructor that allowed
null DFDs as parent.
Storage/DatabaseStorage.php: Made link and node save themselves into the
mapping between DFD and elements - this breaks many tests as many do not
construct a DFD to get an id before saving the nodes/links
Storage/WriteStorable.php: Added parameter to saveLink and saveNode for
parent DFD id