Summary - Node working and Tested
Modified:
Diagram.php: fixed the loadAssociativeArray function to allow for
missing fields

Element.php: updated comments

Link.php:updated comments, fixed constructor to check for both Diagram
and Link IDs and throw an exception if it wasnt one of those.  The
constructor now also adds itself to its parent Diagram when it is
constructed with that specified. Modified the attributes
origin/destination node be associative arrays that contain the id,label
and position of the Node.

Node.php: fixed the constructor to check for Diagram/Node IDs and throw
an exception if that wasnt the case.  The constructor now also adds
itself to its parent Diagram when it is constructed with that specified.
 addLink() can only be called from the Link class now.  updated the
linkList to be a list of associativeArrays that contain the id and label
of each link. Updated the text in several exceptions to be more
descriptive.  removeAllLinks now actually ittereates from the end of the
list to the front of its list

EntityTest.php:modified the setUp() function to not add the object to
the testDiagram (this is now done automatically by the constructor)

NodeTest.php:modified the setUp() function to not add the object to the
testDiagram (this is now done automatically by the constructor).  tests
updated to use the new Constuctor formats.  added a test to ensure that
addLink cannot be called from any class that is not Link.  updated tests
to work with the associative arrays

SimpleMediaType.txt:updated to show the contents of the associative
arrays

DatabaseStorage.php: LoadLink no longer requires the origin and
destination nodes to be set.  Fixed saveNode() to work with the
associative arrays