summary - added updates to the functions that modify the classes in the DiaNode class and the Diagram Class
Diagram.php: added updates to any functions that change a Diagram
object.  changed setParentDiaNode to be protected and it will now also
set itself as that DiaNodes subDiagram.  this function will also throw
an exception when it is passed an Id that did not belong to a dianode
DiaNode: added update() when the DiaNode is modified