Summary - Fixed bug where Links did not manage to remove themeselves from Nodes on delete
Link.php: Correctly handles changes to Nodes it is linked to
Node.php: No longer calls its own updated, handled by Link. Otherwise
this was saving itself, then because another copy of the same Node was
owned by the calling Link - which updated that Node - the Node with the
Link simply got rewritten to storage.