Summary - Updated classloaders to check for a global path
Classloaders need to be able to point to the root directory of DEdC. In
normal cases, this isn't a problem, but PHPUnit doesn't always get that
directory, so an absolute path is required when loading classes, this
change handles that. This path does not have to be set, so the
Controller doesn't need to worry about it.